requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
gunicorn==20.1.0
sqlparse==0.4.2
whitenoise==6.2.0
djangorestframework==3.13.1